+++ Thunderstore/CHANGELOG.md
@@ -19,3 +19,11 @@
 - Created a SCP966 manager which will handle all SCP966 instance in the scene
 ## 1.1.3
 - Fixed a minor problem, where scanning would display an error message if SCP966 was not present in the scene
+## 1.1.4
+- Fixed a problem when destroying  ( NOT KILLING ) an SCP966 instance where he would stay in the SCP manager
+- Made the scan more consistent
+- Changed the curve spawn rate!
+- Put up the base rarity
+- Added configurations!
+- CONFIG : Time you see the monster --> The amount of time the monster is visible on scan
+- CONFIG : Rarity --> The rarity of SCp 966 on all the moons